test fix ups
