+++ apps/mock_forum/test/factories/category_factory.ex
@@ -1,13 +1,13 @@
 defmodule MockForum.CategoryFactory do
+    @moduledoc false
 
-  @moduledoc false
-  alias Faker.Company
-
+    alias Faker.Company
   defmacro __using__(_opts) do
     quote do
       def category_factory do
         %MockForum.Category{
-          title: Company.catch_phrase
+          title: Company.name,
+          description: Company.catch_phrase
         }
       end
     end
