--- apps/mock_forum/test/factories/subject_factory.ex
@@ -1,16 +1,0 @@
-defmodule MockForum.SubjectFactory do
-    @moduledoc false
-
-    alias Faker.Company
-  defmacro __using__(_opts) do
-    quote do
-      def subject_factory do
-        %MockForum.Subject{
-          title: Company.name,
-          description: Company.catch_phrase,
-          category: build(:category)
-        }
-      end
-    end
-  end
-end
