+++ apps/mock_forum/test/support/factories.ex
@@ -2,9 +2,8 @@ defmodule MockForum.Factory do
   @moduledoc false
   # with Ecto
   use ExMachina.Ecto, repo: MockForum.Repo
-  use MockForum.SubjectFactory
+  use MockForum.CategoryFactory
   use MockForum.ThreadFactory
   use MockForum.PostFactory
   use MockForum.UserFactory
-  use MockForum.CategoryFactory
 end
