+++ apps/mock_forum_web/test/mock_forum_web/feature/category_feature_test.exs
@@ -14,35 +14,33 @@ defmodule MockForum.Web.Feature.SubjectFeatureTest do
     {:ok, session: session, user: user}
   end
 
-  describe "Creating a new subject" do
-    test "User can create a new subject and view it", %{session: session} do
-       insert(:category)
-
+  describe "Creating a new category" do
+    test "User can create a new category and view it", %{session: session} do
       session
       |> visit("/")
-      |> click(link("new subject"))
-      |> fill_in(text_field("subject_title"), with: "My Title")
+      |> click(link("new category"))
+      |> fill_in(text_field("category_title"), with: "My Title")
       |> click(button("Submit"))
       |> assert_has(link("My Title"))
     end
 
-    test "A subject name cannot be blank", %{session: session} do
+    test "A category title cannot be blank", %{session: session} do
        insert(:category)
 
       session
       |> visit("/")
-      |> click(link("new subject"))
+      |> click(link("new category"))
       |> click(button("Submit"))
       |> assert_has(css(".help-block", text: "can't be blank"))
     end
   end
 
-  describe "Viewing a new subject" do
-    test "A subject can be clicked on linking to topics", %{session: session} do
-      subject = insert(:subject)
+  describe "Viewing a new category" do
+    test "A category can be clicked on linking to topics", %{session: session} do
+      category = insert(:category)
       session
       |> visit("/")
-      |> click(link(subject.title))
+      |> click(link(category.title))
       |> assert_has(link("New Thread"))
     end
   end
